3.4.1
1. 升级：ZXing3.4.1
2. 升级：Java1.8

+++ src/com/google/zxing/pdf417/encoder/PDF417.java
@@ -702,7 +702,7 @@ public final class PDF417 {
         continue;
       }
 
-      float newRatio = ((17 * cols + 69) * DEFAULT_MODULE_WIDTH) / (rows * HEIGHT);
+      float newRatio = ((float) (17 * cols + 69) * DEFAULT_MODULE_WIDTH) / (rows * HEIGHT);
 
       // ignore if previous ratio is closer to preferred ratio
       if (dimension != null && Math.abs(newRatio - PREFERRED_RATIO) > Math.abs(ratio - PREFERRED_RATIO)) {
@@ -713,13 +713,13 @@ public final class PDF417 {
       dimension = new int[] {cols, rows};
     }
 
-     // Handle case when min values were larger than necessary
-     if (dimension == null) {
-       int rows = calculateNumberOfRows(sourceCodeWords, errorCorrectionCodeWords, minCols);
-       if (rows < minRows) {
-         dimension = new int[]{minCols, minRows};
-       }
-     }
+    // Handle case when min values were larger than necessary
+    if (dimension == null) {
+      int rows = calculateNumberOfRows(sourceCodeWords, errorCorrectionCodeWords, minCols);
+      if (rows < minRows) {
+        dimension = new int[]{minCols, minRows};
+      }
+    }
 
     if (dimension == null) {
       throw new WriterException("Unable to fit message in columns");
